Close unclosed --as-needed linker flag from MySQL.
'mysql_config --libs' emits a -Wl,--as-needed flag but does not close it
(with -Wl,--no-as-needed, which restores the default). This is poor
practice but until now had no negative effect; however, now it removes
libraries that are necessary to create libxbmc. The fix is to add the
restore flag if MYSQL_LIBS contains the flag without a restore flag.