Fix failure to catch exception.
Python loads xbmc.lib's __init__.so with RTLD_LOCAL, which does not make
its symbols available to other modules. When xbmc.lib.scraper is loaded,
scraper.so prefers its own copy of the weak symbols/objects ("W" or "V"
in nm output) for CScraperError and its typeinfo, which is used to match
exceptions. So we wrap the former __init__.so as xbmc.lib._load (and
_load.so) and ensure RTLD_GLOBAL is set before loading it in __init__.py.